+++ libs/community/langchain_community/vectorstores/tencentvectordb.py
@@ -110,11 +110,11 @@ def translate_filter(
     lc_filter: str, allowed_fields: Optional[Sequence[str]] = None
 ) -> str:
     from langchain.chains.query_constructor.base import fix_filter_directive
-    from langchain.chains.query_constructor.ir import FilterDirective
     from langchain.chains.query_constructor.parser import get_parser
     from langchain.retrievers.self_query.tencentvectordb import (
         TencentVectorDBTranslator,
     )
+    from langchain_core.structured_query import FilterDirective
 
     tvdb_visitor = TencentVectorDBTranslator(allowed_fields)
     flt = cast(
