+++ libs/langchain/langchain/retrievers/self_query/astradb.py
@@ -1,7 +1,7 @@
 """Logic for converting internal query language to a valid AstraDB query."""
 from typing import Dict, Tuple, Union
 
-from langchain.chains.query_constructor.ir import (
+from langchain_core.structured_query import (
     Comparator,
     Comparison,
     Operation,
