+++ libs/langchain/langchain/retrievers/self_query/databricks_vector_search.py
@@ -2,7 +2,7 @@ from collections import ChainMap
 from itertools import chain
 from typing import Dict, Tuple
 
-from langchain.chains.query_constructor.ir import (
+from langchain_core.structured_query import (
     Comparator,
     Comparison,
     Operation,
