+++ libs/langchain/langchain/retrievers/self_query/deeplake.py
@@ -1,7 +1,7 @@
 """Logic for converting internal query language to a valid Chroma query."""
 from typing import Tuple, Union
 
-from langchain.chains.query_constructor.ir import (
+from langchain_core.structured_query import (
     Comparator,
     Comparison,
     Operation,
