+++ libs/langchain/langchain/retrievers/self_query/myscale.py
@@ -1,7 +1,7 @@
 import re
 from typing import Any, Callable, Dict, Tuple
 
-from langchain.chains.query_constructor.ir import (
+from langchain_core.structured_query import (
     Comparator,
     Comparison,
     Operation,
