+++ libs/langchain/langchain/retrievers/self_query/vectara.py
@@ -1,6 +1,6 @@
 from typing import Tuple, Union
 
-from langchain.chains.query_constructor.ir import (
+from langchain_core.structured_query import (
     Comparator,
     Comparison,
     Operation,
