+++ libs/langchain/langchain/retrievers/self_query/weaviate.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 from typing import Dict, Tuple, Union
 
-from langchain.chains.query_constructor.ir import (
+from langchain_core.structured_query import (
     Comparator,
     Comparison,
     Operation,
