+++ libs/langchain/tests/unit_tests/chains/query_constructor/test_parser.py
@@ -3,13 +3,13 @@ from typing import Any, cast
 
 import lark
 import pytest
-
-from langchain.chains.query_constructor.ir import (
+from langchain_core.structured_query import (
     Comparator,
     Comparison,
     Operation,
     Operator,
 )
+
 from langchain.chains.query_constructor.parser import get_parser
 
 DEFAULT_PARSER = get_parser()
