+++ libs/langchain/tests/unit_tests/retrievers/self_query/test_chroma.py
@@ -1,12 +1,13 @@
 from typing import Dict, Tuple
 
-from langchain.chains.query_constructor.ir import (
+from langchain_core.structured_query import (
     Comparator,
     Comparison,
     Operation,
     Operator,
     StructuredQuery,
 )
+
 from langchain.retrievers.self_query.chroma import ChromaTranslator
 
 DEFAULT_TRANSLATOR = ChromaTranslator()
