+++ libs/langchain/tests/unit_tests/retrievers/self_query/test_dashvector.py
@@ -1,13 +1,13 @@
 from typing import Any, Tuple
 
 import pytest
-
-from langchain.chains.query_constructor.ir import (
+from langchain_core.structured_query import (
     Comparator,
     Comparison,
     Operation,
     Operator,
 )
+
 from langchain.retrievers.self_query.dashvector import DashvectorTranslator
 
 DEFAULT_TRANSLATOR = DashvectorTranslator()
