+++ libs/langchain/tests/unit_tests/retrievers/self_query/test_databricks_vector_search.py
@@ -1,14 +1,14 @@
 from typing import Any, Dict, Tuple
 
 import pytest
-
-from langchain.chains.query_constructor.ir import (
+from langchain_core.structured_query import (
     Comparator,
     Comparison,
     Operation,
     Operator,
     StructuredQuery,
 )
+
 from langchain.retrievers.self_query.databricks_vector_search import (
     DatabricksVectorSearchTranslator,
 )
