+++ libs/langchain/tests/unit_tests/retrievers/self_query/test_milvus.py
@@ -1,14 +1,14 @@
 from typing import Any, Dict, Tuple
 
 import pytest
-
-from langchain.chains.query_constructor.ir import (
+from langchain_core.structured_query import (
     Comparator,
     Comparison,
     Operation,
     Operator,
     StructuredQuery,
 )
+
 from langchain.retrievers.self_query.milvus import MilvusTranslator
 
 DEFAULT_TRANSLATOR = MilvusTranslator()
