+++ libs/langchain/tests/unit_tests/retrievers/self_query/test_myscale.py
@@ -1,14 +1,14 @@
 from typing import Any, Dict, Tuple
 
 import pytest
-
-from langchain.chains.query_constructor.ir import (
+from langchain_core.structured_query import (
     Comparator,
     Comparison,
     Operation,
     Operator,
     StructuredQuery,
 )
+
 from langchain.retrievers.self_query.myscale import MyScaleTranslator
 
 DEFAULT_TRANSLATOR = MyScaleTranslator()
