+++ libs/langchain/tests/unit_tests/retrievers/self_query/test_tencentvectordb.py
@@ -1,10 +1,11 @@
-from langchain.chains.query_constructor.ir import (
+from langchain_core.structured_query import (
     Comparator,
     Comparison,
     Operation,
     Operator,
     StructuredQuery,
 )
+
 from langchain.retrievers.self_query.tencentvectordb import TencentVectorDBTranslator
