+++ libs/langchain/tests/unit_tests/retrievers/self_query/test_timescalevector.py
@@ -1,14 +1,14 @@
 from typing import Dict, Tuple
 
 import pytest as pytest
-
-from langchain.chains.query_constructor.ir import (
+from langchain_core.structured_query import (
     Comparator,
     Comparison,
     Operation,
     Operator,
     StructuredQuery,
 )
+
 from langchain.retrievers.self_query.timescalevector import TimescaleVectorTranslator
 
 DEFAULT_TRANSLATOR = TimescaleVectorTranslator()
